Dostęp do zasobu per user

+++ RestaurantAPI/RestaurantAPI/Services/RestaurantService.cs
@@ -1,10 +1,14 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantAPI.Authorization;
+using RestaurantAPI.Enums;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Interfaces;
 using RestaurantAPI.Model.DatabaseConnection;
 using RestaurantAPI.Model.Entities;
 using RestaurantAPI.Model.Models;
+using System.Security.Claims;
 
 namespace RestaurantAPI.Services
 {
@@ -13,15 +17,17 @@ namespace RestaurantAPI.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _authorizationService;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger, IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
 
-        public void Delete(int id)
+        public void Delete(int id, ClaimsPrincipal user)
         {
             _logger.LogWarning($"Restauracja z ID {id} wywołanie metody DELETE");
             var restaurant = _dbContext
@@ -29,6 +35,14 @@ namespace RestaurantAPI.Services
                 .FirstOrDefault(x => x.Id == id);
 
             if (restaurant == null) throw new NotFoundException($"Nie znaleziono zasobu o ID: {id}");
+
+            var autherizationRes = _authorizationService.AuthorizeAsync(user, restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!autherizationRes.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
@@ -62,11 +76,12 @@ namespace RestaurantAPI.Services
             return restaurantsDto;
         }
 
-        public int Create(CreateRestaurantDto dto)
+        public int Create(CreateRestaurantDto dto, int userId)
         {
             _logger.LogWarning($"Wywołanie metody Create");
 
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = userId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
 
@@ -75,7 +90,7 @@ namespace RestaurantAPI.Services
             return restaurant.Id;
         }
 
-        public void Update(UpdateRestaurantDto resraurantRequest, int id)
+        public void Update(UpdateRestaurantDto resraurantRequest, int id, ClaimsPrincipal user)
         {
             _logger.LogWarning($"Wywołanie metody Update dla restauracji {id}");
 
@@ -84,6 +99,14 @@ namespace RestaurantAPI.Services
                 .FirstOrDefault(x => x.Id == id);
 
             if (restaurant == null) throw new NotFoundException($"Nie znaleziono zasobu o ID: {id}");
+
+            var autherizationRes = _authorizationService.AuthorizeAsync(user, restaurant, 
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!autherizationRes.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             restaurant.UpdateFrom(resraurantRequest);
             _dbContext .SaveChanges();
